flows: add test for diagram
